feat: added format time functions
